Refactor main command initialization by modularizing commands and moving them into dedicated packages. Simplify `rootCmd` setup and reconfigure persistent pre-run logic.
