+++ cmd/clock/clock.go
@@ -5,7 +5,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-func Cmd(client *odoo.XMLRPCClient) *cobra.Command {
+func CMD(client *odoo.XMLRPCClient) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "clock",
 		Short: "Clock in/out and attendance status",
